+++ PhzQtUI/src/lib/GridButton.cpp
@@ -11,7 +11,8 @@ namespace Euclid {
 namespace PhzQtUI {
 
 
-  GridButton::GridButton(int x, int y, const QString &text, QWidget *parent) : QPushButton(text,parent),m_x{x},m_y{y}{
+  GridButton::GridButton(size_t x, size_t y, const QString &text, QWidget *parent)
+  : QPushButton(text,parent),m_x{x},m_y{y}{
     QObject::connect( this, SIGNAL(clicked()), this,SLOT(recievClicked()));
   }
